Updated Bookmark and Endorsement models to work with new DB structure.

+++ gnosis/catalog/views/views_endorsement.py
@@ -2,13 +2,13 @@ from django.contrib.auth.decorators import login_required
 from django.contrib.admin.views.decorators import staff_member_required
 from django.shortcuts import render, get_object_or_404
 from django.http import Http404
-from catalog.models import EndorsementEntry
+from catalog.models import Endorsement, Paper
 from django.urls import reverse
 from django.http import HttpResponseRedirect
-#from catalog.views.views import _get_paper_by_id
 #
 # Endorsement views
 #
+
 
 @login_required
 def endorsements(request):
@@ -18,46 +18,48 @@ def endorsements(request):
     return render(request, "endorsement.html", {"papers": endorsement, })
 
 
-
 @login_required
-def endorsement_create(request, paper_id):
+def endorsement_create(request, id):
     """Create an endorsement entry of a paper by the user, update the database"""
     user = request.user
+    paper = get_object_or_404(Paper, pk=id)
 
     if request.method == "POST":
-        e = EndorsementEntry.objects.filter(paper_id = paper_id).filter(user = user)
+        e = Endorsement.objects.filter(paper=paper).filter(user=user)
         if not e:
-            endorsement_entry = EndorsementEntry()
+            endorsement_entry = Endorsement()
             endorsement_entry.user = user
-            endorsement_entry.paper_id = paper_id
-            endorsement_entry.paper_title = ""  # _get_paper_by_id(paper_id).title
+            endorsement_entry.paper = paper
             endorsement_entry.save()
 
-    return HttpResponseRedirect(reverse("paper_detail", kwargs={"id": paper_id}))
+    return HttpResponseRedirect(reverse("paper_detail", kwargs={"id": id}))
 
 
 @login_required
-def endorsement_undo(request, paper_id):
+def endorsement_undo(request, id):
     """Undo an endorsement of a paper, update both databases"""
     user = request.user
+    paper = get_object_or_404(Paper, pk=id)
+
     if request.method == "POST":
         try:
-            endorsement_entry = EndorsementEntry.objects.filter(paper_id=paper_id, user=user)[0]
-            endorsement_entry.delete()
+            endorsement_entry = Endorsement.objects.filter(paper=paper, user=user)
+            if endorsement_entry:
+                endorsement_entry.delete()
         except:
             print("no such entry exists")
-        return HttpResponseRedirect(reverse("paper_detail", kwargs={"id": paper_id}))
+    return HttpResponseRedirect(reverse("paper_detail", kwargs={"id": id}))
 
-    return render(request, "paper_detail.html")
 
 @login_required
-def endorsement_undo_from_view(request, paper_id):
+def endorsement_undo_from_view(request, id):
     """Undo an endorsement of a paper, update both databases"""
     user = request.user
+    paper = get_object_or_404(Paper, pk=id)
     try:
-        endorsement_entry = EndorsementEntry.objects.filter(paper_id=paper_id, user=user)[0]
+        endorsement_entry = Endorsement.objects.filter(paper=paper, user=user)
         endorsement_entry.delete()
-        message = None
     except:
-        message = "no such entry exists"
+        print("Endorsment not found!")
+
     return HttpResponseRedirect(reverse("endorsements"))
